+++ codec/src/codec/encode.rs
@@ -4,11 +4,11 @@ use chrono::{DateTime, Utc};
 use ntex_bytes::{BufMut, ByteString, Bytes, BytesMut};
 use uuid::Uuid;
 
-use crate::codec::{self, ArrayEncode, Encode};
+use crate::codec::{self, ArrayEncode, Composite, Encode};
 use crate::framing::{self, AmqpFrame, SaslFrame};
 use crate::types::{
-    Constructor, Descriptor, List, Multiple, StaticSymbol, Str, Symbol, Variant, VecStringMap,
-    VecSymbolMap,
+    Constructor, Descriptor, List, ListDescribed, Multiple, StaticSymbol, Str, Symbol, Variant,
+    VecStringMap, VecSymbolMap,
 };
 
 fn encode_null(buf: &mut BytesMut) {
@@ -625,13 +625,9 @@ impl<T: Encode + ArrayEncode> Encode for Multiple<T> {
     }
 }
 
-fn list_encoded_size(vec: &List) -> usize {
-    vec.iter().fold(0, |r, i| r + i.encoded_size())
-}
-
 impl Encode for List {
     fn encoded_size(&self) -> usize {
-        let content_size = list_encoded_size(self);
+        let content_size = self.iter().fold(0, |r, i| r + i.encoded_size());
         // format_code + size + count
         (if content_size + 1 > u8::MAX as usize {
             9
@@ -641,7 +637,7 @@ impl Encode for List {
     }
 
     fn encode(&self, buf: &mut BytesMut) {
-        let size = list_encoded_size(self);
+        let size = self.iter().fold(0, |r, i| r + i.encoded_size());
         if size + 1 > u8::MAX as usize {
             buf.put_u8(codec::FORMATCODE_LIST32);
             buf.put_u32((size + 4) as u32); // +4 for 4 byte count that follow
@@ -652,6 +648,44 @@ impl Encode for List {
             buf.put_u8(self.len() as u8);
         }
         for i in self.iter() {
+            i.encode(buf);
+        }
+    }
+}
+
+impl<T: Composite> Encode for ListDescribed<T> {
+    fn encoded_size(&self) -> usize {
+        let descr_size = T::descriptor().encoded_size();
+        let content_size = self
+            .iter()
+            .fold(0, |r, i| r + i.encoded_size() + descr_size);
+
+        // format_code + size + count
+        (if content_size + 1 > u8::MAX as usize {
+            9
+        } else {
+            3
+        }) + content_size
+    }
+
+    fn encode(&self, buf: &mut BytesMut) {
+        let descr = T::descriptor();
+        let descr_size = descr.encoded_size();
+        let content_size = self
+            .iter()
+            .fold(0, |r, i| r + i.encoded_size() + descr_size);
+
+        if content_size + 1 > u8::MAX as usize {
+            buf.put_u8(codec::FORMATCODE_LIST32);
+            buf.put_u32((content_size + 4) as u32); // +4 for 4 byte count that follow
+            buf.put_u32(self.len() as u32);
+        } else {
+            buf.put_u8(codec::FORMATCODE_LIST8);
+            buf.put_u8((content_size + 1) as u8); // +1 for 1 byte count that follow
+            buf.put_u8(self.len() as u8);
+        }
+        for i in self.iter() {
+            descr.encode(buf);
             i.encode(buf);
         }
     }